Move WrapperFor enum to constant file
