+++ src/resolver/constant.ts
@@ -1,0 +1,5 @@
+export enum WrapperFor {
+  TYPE = 'TYPE',
+  FIELD = 'FIELD',
+  ANY = 'ANY',
+}
