+++ src/types.ts
@@ -29,35 +29,6 @@ export type ResolverInfo = Parameters<GraphQLFieldResolver<any, any>>[3];
 
 // Library Abstractions
 
-export type TypeWrapperFunction = (
-  resolver: TypeResolver,
-  options: WrapperOptionsBase,
-) => TypeResolver | Promise<TypeResolver>;
-
-export type FieldWrapperFunction = (
-  resolver: FieldResolver,
-  options: WrapperOptionsBase,
-) => FieldResolver | Promise<FieldResolver>;
-
-export type GenericWrapperFunction = (
-  resolver: FieldResolver | TypeResolver,
-  options: WrapperOptionsBase,
-) => FieldResolver | TypeResolver | Promise<FieldResolver | TypeResolver>;
-
-export enum WrapperFor {
-  TYPE = 'TYPE',
-  FIELD = 'FIELD',
-  ANY = 'ANY',
-}
-
-export interface NamedWrapper {
-  name: string;
-  wrap: GenericWrapperFunction;
-  wrapperFor: WrapperFor;
-}
-
-export type Wrapper = NamedWrapper | GenericWrapperFunction | FieldWrapperFunction | TypeWrapperFunction;
-
 export type WrapperOptionsBase = {
   schema: GraphQLSchema;
   resolverMap: ResolverMap;
