+++ test/unit/resolver/wrap.test.ts
@@ -3,8 +3,10 @@ import { spy, SinonSpy } from 'sinon';
 import { wrap } from '../../../src/resolver/wrap';
 import { generatePackOptions, userObjectType, userObjectNameField, nameableInterfaceType } from '../../mocks';
 import { GraphQLSchema, GraphQLResolveInfo, GraphQLAbstractType } from 'graphql';
-import { WrapperOptionsBase, FieldResolver, FieldWrapperFunction, WrapperFor, TypeResolver } from '../../../src/types';
+import { WrapperOptionsBase, FieldResolver, TypeResolver } from '../../../src/types';
 import { createWrapper } from '../../../src/resolver/create-wrapper';
+import { FieldWrapperFunction } from '../../../src/resolver/types';
+import { WrapperFor } from '../../../src/resolver/constant';
 
 describe('resolver/wrap', function () {
   let resolverWrapperOptions: WrapperOptionsBase;
